Basic follower/following support
Woo! This was quite a pain in the ass considering I tried to
change things up a bit. I wanted to use a Follower model, but
I didn't snap that I was going to get a follower_id that would
be the primary key. All of the code from the tutorial assumed
that there was a column follower_id that was not a primary key.

I finally figured out the problem and fixed up the code. I also
had a couple of typos in the relationship model.